Pass redux store by Provider. Use mAction to call rAction, dispatch to redux state.
important:
TODO: onClick function not finished.
