+++ client/main.js
@@ -1,6 +1,7 @@
 import { createStore, combineReducers, compose } from 'redux';
 import {createApp} from 'mantra-core';
 import initContext from './configs/context';
+import reduxStore from './configs/reduxContext'
 // INSERT IMPORT
 
 // modules
@@ -8,20 +9,9 @@ import coreModule from "./modules/core";
 import bpMeasureModule from "./modules/bpMeasure";
 import tableDemoModule from "./modules/tableDemo";
 
-const reducer = combineReducers({
-  ...coreModule.reducers,
-  ...bpMeasureModule.reducers
-});
-
-
-function configureStore() {
-    const store = createStore(reducer, compose(
-      window.devToolsExtension ? window.devToolsExtension() : f => f
-    ));
-    return store;
-  }
 // init context
-const context = initContext(configureStore());
+// const context = initContext(reduxStore());  //尝试抛弃mantra传递redux store
+const context = initContext()
 
 // create app
 const app = createApp(context);
